+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switch to Rust 2024 edition.
 - Switched internal data store to slot storage.
 - Syntax for parameters has changed from `{name}` to `<name>`
+- Parameter characters `<` and `>` can no longer be escaped.
 
 ### Fixed